+++ bonus/library/Libft_vector/matrix44.h
@@ -6,7 +6,7 @@
 /*   By: dkim2 <dkim2@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/11 12:17:35 by dkim2             #+#    #+#             */
-/*   Updated: 2022/08/16 08:48:57 by dkim2            ###   ########.fr       */
+/*   Updated: 2022/08/16 17:52:05 by dkim2            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ t_mat44	trans_mat44(t_mat44 a);
 
 t_mat44	mul_mat44(t_mat44 a, t_mat44 b);
 
-float	quadratic_form(t_vec4 v1, t_mat44 m, t_vec4 v2);
+double	quadratic_form(t_vec4 v1, t_mat44 m, t_vec4 v2);
 
 t_mat44	rotation_mat44(t_vec3 axis, double deg);
