+++ bonus/ray_cast/diffuse_reflection.c
@@ -6,7 +6,7 @@
 /*   By: dkim2 <dkim2@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/06 15:22:41 by dkim2             #+#    #+#             */
-/*   Updated: 2022/08/16 12:04:49 by dkim2            ###   ########.fr       */
+/*   Updated: 2022/08/16 17:53:36 by dkim2            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 #include <stdlib.h>
 #include "ray_cast.h"
 
-static float	diffuse_helper(t_quadrics *objlst, \
+static double	diffuse_helper(t_quadrics *objlst, \
 							t_light *target_light, \
 							t_vec3 normal, \
 							t_vec3 hit_point)
@@ -53,7 +53,7 @@ t_color	diffuse_light(const t_scene *scene, \
 	t_color	color;
 	t_color	clr_tmp;
 	t_color	obj_color;
-	float	diffuse;
+	double	diffuse;
 
 	hit_point = v3_add(hit_point, v3_mul(normal, EPSILON));
 	color = rgb_color(0, 0, 0);
@@ -65,10 +65,10 @@ t_color	diffuse_light(const t_scene *scene, \
 		diffuse = diffuse_helper(scene->quads, light, normal, hit_point);
 		if (diffuse > EPSILON)
 		{
-			clr_tmp.red = (int)((float)light->color.red / 255 * obj_color.red);
-			clr_tmp.green = (int)((float)light->color.green / 255 
+			clr_tmp.red = (int)((double)light->color.red / 255 * obj_color.red);
+			clr_tmp.green = (int)((double)light->color.green / 255 
 									* obj_color.green);
-			clr_tmp.blue = (int)((float)light->color.blue / 255 * obj_color.blue);
+			clr_tmp.blue = (int)((double)light->color.blue / 255 * obj_color.blue);
 			clr_tmp = color_scale(clr_tmp, diffuse * (1 - Q->spec_rs));
 			clr_tmp = color_scale(clr_tmp, diffuse);
 			color = color_add(color, clr_tmp);
