+++ bonus/ray_cast/ray_cast.h
@@ -6,7 +6,7 @@
 /*   By: dkim2 <dkim2@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/01 16:07:19 by dkim2             #+#    #+#             */
-/*   Updated: 2022/08/12 12:21:54 by dkim2            ###   ########.fr       */
+/*   Updated: 2022/08/16 17:53:39 by dkim2            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ typedef struct s_thread_local_object
 	int		y;
 }			t_thread_local_object;
 
-float	get_intersect_distance(t_quadrics *objlst, \
+double	get_intersect_distance(t_quadrics *objlst, \
 								t_quadrics **intersecting_obj_out, \
 								t_ray ray);
